Minor fixes to gaussian2d jacobian.
Transpose the output, so we're effectively giving a list of
jacobians, one for an evaluation at each x,y pair. This matches
the behaviour of the plain gaussian2d function, and seems to
give the shape of array expected by least_squares.

Unfortunately, employing the jacobian (in the current tests)
doesn't seem to work well - most fits are much worse to the point
of being failures, even though it supposedly converged.